+++ src/otx/engine/engine.py
@@ -5,9 +5,11 @@
 
 from __future__ import annotations
 
+import csv
 import inspect
 import logging
 import tempfile
+import time
 from contextlib import contextmanager
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, ClassVar, Iterable, Iterator, Literal
@@ -28,7 +30,7 @@ from otx.core.types.export import OTXExportFormatType
 from otx.core.types.precision import OTXPrecisionType
 from otx.core.types.task import OTXTaskType
 from otx.core.utils.cache import TrainerArgumentsCache
-from otx.utils.utils import is_xpu_available
+from otx.utils.utils import is_xpu_available, measure_flops
 
 from .adaptive_bs import adapt_batch_size
 from .hpo import execute_hpo, update_hyper_parameter
@@ -119,6 +121,7 @@ class Engine:
         model: OTXModel | str | None = None,
         checkpoint: PathLike | None = None,
         device: DeviceType = DeviceType.auto,
+        num_devices: int = 1,
         **kwargs,
     ):
         """Initializes the OTX Engine.
@@ -131,12 +134,14 @@ class Engine:
             model (OTXModel | str | None, optional): The model for the engine. Defaults to None.
             checkpoint (PathLike | None, optional): Path to the checkpoint file. Defaults to None.
             device (DeviceType, optional): The device type to use. Defaults to DeviceType.auto.
+            num_devices (int, optional): The number of devices to use. If it is 2 or more, it will behave as multi-gpu.
             **kwargs: Additional keyword arguments for pl.Trainer.
         """
         self._cache = TrainerArgumentsCache(**kwargs)
         self.checkpoint = checkpoint
         self.work_dir = work_dir
         self.device = device  # type: ignore[assignment]
+        self.num_devices = num_devices
         self._auto_configurator = AutoConfigurator(
             data_root=data_root,
             task=datamodule.task if datamodule is not None else task,
@@ -561,7 +566,7 @@ class Engine:
             export_demo_package = False
 
         if is_ir_ckpt and not export_demo_package:
-            msg = "IR model is passed as a checkpoint, export automaticaly switched to exportable code."
+            msg = "IR model is passed as a checkpoint, export automatically switched to exportable code."
             warn(msg, stacklevel=1)
             export_demo_package = True
 
@@ -778,6 +783,140 @@ class Engine:
         model.explain_mode = False
         return predict_result
 
+    def benchmark(
+        self,
+        checkpoint: PathLike | None = None,
+        batch_size: int = 1,
+        n_iters: int = 10,
+        extended_stats: bool = False,
+        print_table: bool = True,
+    ) -> dict[str, str]:
+        r"""Executes model micro benchmarking on random data.
+
+        Benchmark can provide latency, throughput, number of parameters,
+        and theoretical computational complexity with batch size 1.
+        The latter two characteristics are available for torch model recipes only.
+        Before the measurements, a warm-up is done.
+
+        Args:
+            checkpoint (PathLike | None, optional): Path to checkpoint. Optional for torch models. Defaults to None.
+            batch_size (int, optional): Batch size for benchmarking. Defaults to 1.
+            n_iters (int, optional): Number of iterations to average on. Defaults to 10.
+            extended_stats (bool, optional): Flag that enables printing of per module complexity for torch model.
+                Defaults to False.
+            print_table (bool, optional): Flag that enables printing the benchmark results in a rich table.
+                Defaults to True.
+
+        Returns:
+            dict[str, str]: a dict with the benchmark results.
+
+        Example:
+            >>> engine.benchmark(
+            ...     checkpoint=<checkpoint-path>,
+            ...     batch_size=1,
+            ...     n_iters=20,
+            ...     extended_stats=True,
+            ... )
+
+        CLI Usage:
+            1. To run benchmark by specifying the work_dir where did the training, run
+                ```shell
+                >>> otx benchmark --work_dir <WORK_DIR_PATH, str>
+                ```
+            2. To run benchmark by specifying the checkpoint, run
+                ```shell
+                >>> otx benchmark \
+                ...     --work_dir <WORK_DIR_PATH, str> \
+                ...     --checkpoint <CKPT_PATH, str>
+                ```
+            3. To run benchmark using the configuration, launch
+                ```shell
+                >>> otx benchmark \
+                ...     --config <CONFIG_PATH> \
+                ...     --data_root <DATASET_PATH, str> \
+                ...     --checkpoint <CKPT_PATH, str>
+                ```
+        """
+        checkpoint = checkpoint if checkpoint is not None else self.checkpoint
+
+        if checkpoint is not None:
+            is_ir_ckpt = Path(checkpoint).suffix in [".xml"]
+            if is_ir_ckpt and not isinstance(self.model, OVModel):
+                # create OVModel
+                self.model = self._auto_configurator.get_ov_model(
+                    model_name=str(checkpoint),
+                    label_info=self.datamodule.label_info,
+                )
+
+            if not is_ir_ckpt:
+                model_cls = self.model.__class__
+                self.model = model_cls.load_from_checkpoint(
+                    checkpoint_path=checkpoint,
+                    map_location="cpu",
+                    **self.model.hparams,
+                )
+        elif isinstance(self.model, OVModel):
+            msg = "To run benchmark on OV model, checkpoint must be specified."
+            raise RuntimeError(msg)
+
+        self.model.eval()
+
+        def dummy_infer(model: OTXModel, batch_size: int = 1) -> float:
+            input_batch = model.get_dummy_input(batch_size)
+            start = time.perf_counter()
+            model.forward(input_batch)
+            end = time.perf_counter()
+            return end - start
+
+        warmup_iters = max(1, int(n_iters / 10))
+        for _ in range(warmup_iters):
+            dummy_infer(self.model, batch_size)
+
+        total_time = 0.0
+        for _ in range(n_iters):
+            total_time += dummy_infer(self.model, batch_size)
+        latency = total_time / n_iters
+        fps = batch_size / latency
+
+        final_stats = {"latency": f"{latency:.3f} s", "throughput": f"{(fps):.3f} FPS"}
+
+        if not isinstance(self.model, OVModel):
+            try:
+                from torch.utils.flop_counter import convert_num_with_suffix, get_suffix_str
+
+                input_batch = self.model.get_dummy_input(1)
+                model_fwd = lambda: self.model.forward(input_batch)
+                depth = 3 if extended_stats else 0
+                fwd_flops = measure_flops(self.model.model, model_fwd, print_stats_depth=depth)
+                flops_str = convert_num_with_suffix(fwd_flops, get_suffix_str(fwd_flops * 10**3))
+                final_stats["complexity"] = flops_str + " MACs"
+            except Exception as e:
+                logging.warning(f"Failed to complete complexity estimation: {e}")
+
+            params_num = sum(p.numel() for p in self.model.parameters() if p.requires_grad)
+            params_num_str = convert_num_with_suffix(params_num, get_suffix_str(params_num * 100))
+            final_stats["parameters_number"] = params_num_str
+
+        if print_table:
+            from rich.console import Console
+            from rich.table import Column, Table
+
+            console = Console()
+            table_headers = ["Benchmark", "Value"]
+            columns = [Column(h, justify="center", style="magenta", width=console.width) for h in table_headers]
+            columns[0].style = "cyan"
+            table = Table(*columns)
+            for name, val in final_stats.items():
+                table.add_row(*[f"{name:<20}", f"{val}"])
+            console.print(table)
+
+        with (Path(self.work_dir) / "benchmark_report.csv").open("w") as f:
+            writer = csv.writer(f)
+            writer.writerow(list(final_stats))
+            writer.writerow(list(final_stats.values()))
+
+        return final_stats
+
     @classmethod
     def from_config(
         cls,
@@ -944,6 +1083,18 @@ class Engine:
             device = DeviceType.xpu
         self._device = DeviceConfig(accelerator=device)
         self._cache.update(accelerator=self._device.accelerator, devices=self._device.devices)
+        self._cache.is_trainer_args_identical = False
+
+    @property
+    def num_devices(self) -> int:
+        """Number of devices for Engine use."""
+        return self._device.devices
+
+    @num_devices.setter
+    def num_devices(self, num_devices: int) -> None:
+        """Setter function for multi-gpu."""
+        self._device.devices = num_devices
+        self._cache.update(devices=self._device.devices)
         self._cache.is_trainer_args_identical = False
 
     @property