+++ CHANGELOG.md
@@ -18,6 +18,8 @@ All notable changes to this project will be documented in this file.
   (https://github.com/openvinotoolkit/training_extensions/pull/3749)
 - Enable torch.compile to work with classification
   (https://github.com/openvinotoolkit/training_extensions/pull/3758)
+- Add `otx benchmark` subcommand
+  (https://github.com/openvinotoolkit/training_extensions/pull/3762)
 
 ### Enhancements
 
@@ -34,6 +36,8 @@ All notable changes to this project will be documented in this file.
   (https://github.com/openvinotoolkit/training_extensions/pull/3723)
 - Revert #3579 to fix issues with replacing coco_instance with a different format in some dataset
   (https://github.com/openvinotoolkit/training_extensions/pull/3753)
+- Add num_devices in Engine for multi-gpu training
+  (https://github.com/openvinotoolkit/training_extensions/pull/3778)
 
 ## \[v2.1.0\]